medclaim-ai/requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# AI/ML libraries
langchain>=0.1.0
langchain-community>=0.0.13
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-ollama>=0.1.0
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Document processing
PyMuPDF>=1.23.0
docling>=1.0.0
python-docx>=0.8.11

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration and utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
